Module codetrack-annotation is renamed to codetrack-annotation-java
