+++ codetrack-annotation-java/src/main/java/org/codetrack/annotation/Artifacts.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2015 the original author or authors members of codetrack.org
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package org.codetrack.annotation;
+
+import java.lang.annotation.*;
+
+/**
+ * This annotation links external artifacts to the class operation
+ *
+ * @author josecmoj josecmoj@gmail.com
+ */
+@Retention(RetentionPolicy.SOURCE)
+@Target({ElementType.TYPE, ElementType.METHOD})
+@Documented
+public @interface Artifacts {
+
+    /**
+     * Artifacts list
+     *
+     * @return String array of artifacts
+     */
+    String[] artifacts();
+
+}
