--- codetrack-annotation/src/main/java/org/codetrack/annotation/Cases.java
@@ -1,14 +1,0 @@
-package org.codetrack.annotation;
-
-import java.lang.annotation.*;
-
-/**
- * Created by josecmoj on 16/04/15.
- */
-@Retention(RetentionPolicy.SOURCE)
-@Target({ElementType.TYPE})
-@Documented
-public @interface Cases {
-
-    Case[] cases();
-}
