--- codetrack-annotation/src/main/java/org/codetrack/annotation/FixType.java
@@ -1,6 +1,0 @@
-package org.codetrack.annotation;
-
-public enum FixType {
-
-	ALM, EXP;
-}
