--- codetrack-annotation/src/main/java/org/codetrack/annotation/Fixes.java
@@ -1,11 +1,0 @@
-package org.codetrack.annotation;
-
-import java.lang.annotation.*;
-
-@Retention(RetentionPolicy.SOURCE)
-@Target({ElementType.TYPE, ElementType.METHOD})
-@Documented
-public @interface Fixes {
-
-	Fix[] fixes();
-}
