--- codetrack-annotation/src/main/java/org/codetrack/annotation/Rule.java
@@ -1,14 +1,0 @@
-package org.codetrack.annotation;
-
-import java.lang.annotation.*;
-
-@Retention(RetentionPolicy.SOURCE)
-@Target({ElementType.METHOD})
-@Documented
-public @interface Rule {
-
-    String id();
-
-	String caseid();
-
-}
