--- codetrack-annotation/src/main/java/org/codetrack/annotation/Rules.java
@@ -1,11 +1,0 @@
-package org.codetrack.annotation;
-
-import java.lang.annotation.*;
-
-@Retention(RetentionPolicy.SOURCE)
-@Target({ElementType.METHOD})
-@Documented
-public @interface Rules {
-
-	Rule[] rule();
-}
